+++ scripts/build-repo-deb.sh
@@ -6,6 +6,19 @@
 # Sounce config variables
 source config/cli.sh
 
+# Check channel/component/environment
+if [[ "$1" == "production" || "$1" == "canary" || "$1" == "testing" ]]; then
+  if [[ "$1" == "production" ]]; then
+    CHANNEL="main"
+  else
+    CHANNEL=$1
+  fi
+  echo "Building moriod .deb repo package for $1 (channel = $CHANNEL)"
+else
+  echo "Invalid distribution channel. Please specify one of: production, canary, testing."
+  exit 1
+fi
+
 # Create package structure
 cd $MORIO_GIT_ROOT
 rm -rf build-context/*
@@ -15,7 +28,7 @@ mkdir -p build-context/etc/apt/sources.list.d
 mkdir -p build-context/var
 cp morio.gpg build-context/usr/share/keyrings/
 echo "# Moriod repository. See https://apt.repo.morio.it/" > build-context/etc/apt/sources.list.d/moriod.list
-echo "deb [signed-by=/usr/share/keyrings/morio.gpg] https://apt.repo.morio.it/ $(lsb_release -cs) main" >> build-context/etc/apt/sources.list.d/moriod.list
+echo "deb [signed-by=/usr/share/keyrings/morio.gpg] https://apt.repo.morio.it/ $(lsb_release -cs) $CHANNEL" >> build-context/etc/apt/sources.list.d/moriod.list
 cat config/moriod-repos/deb/control | sed "s/MORIO_VERSION/${MORIO_VERSION}/g" > build-context/control
 cp config/moriod-repos/deb/postinst build-context/